Update test execution with results collector function

+++ ci-operator/step-registry/camel-quarkus/execute-tests/camel-quarkus-execute-tests-commands.sh
@@ -6,22 +6,29 @@ set -o pipefail
 
 shopt -s nullglob
 
-echo "Logging to oc:"
-sh ./oc_login.sh
-echo "Executing tests:"
-sh ./run.sh
-
 # Collect test reports into $ARTIFACT_DIR
 PROJECTS=$PROJECTS
 
-echo "Copying results and xmls to ${ARTIFACT_DIR}"
-for PROJECT in ${PROJECTS//","/" "}; do
-  for FILE in ./$PROJECT/target/failsafe-reports/TEST-*.xml; do
-    FILENAME=$(basename $FILE)
-    TARGET=${ARTIFACT_DIR}/junit_${FILENAME}
-    echo "Collecting ${TARGET}"
-    cp $FILE $TARGET
-  done
-done
+function collect-results() {
+    echo "Copying results and xmls to ${ARTIFACT_DIR}"
+    for PROJECT in ${PROJECTS//","/" "}; do
+      for FILE in ./$PROJECT/target/failsafe-reports/TEST-*.xml; do
+        FILENAME=$(basename $FILE)
+        if [[ -f "${FILE}" ]]; then
+            TARGET=${ARTIFACT_DIR}/junit_${FILENAME}
+            echo "Collecting ${TARGET}"
+            cp $FILE $TARGET
+        fi
+      done
+    done
+}
+
+echo "Logging to oc:"
+sh ./oc_login.sh
+echo "Executing tests:"
+trap collect-results SIGINT SIGTERM ERR EXIT
+sh ./run.sh
+
+
